Move tests regarding alias properties of IdP endpoints to separate class
